ISDB-374: Open api configuration Service bus subscription apply on feature branch
