+++ src/main/java/uk/gov/hmcts/reform/hmc/api/Application.java
@@ -8,7 +8,7 @@ import org.springframework.cloud.openfeign.EnableFeignClients;
 import org.springframework.context.annotation.Bean;
 import org.springframework.context.annotation.Import;
 import org.springframework.messaging.Message;
-import uk.gov.hmcts.reform.hmc.api.config.openapi.OpenAPIConfiguration;
+import uk.gov.hmcts.reform.hmc.api.config.OpenAPIConfiguration;
 
 @Slf4j
 @EnableFeignClients(
