--- src/main/java/uk/gov/hmcts/reform/hmc/api/config/openapi/OpenAPIConfiguration.java
@@ -1,25 +1,0 @@
-package uk.gov.hmcts.reform.hmc.api.config.openapi;
-
-import io.swagger.v3.oas.models.ExternalDocumentation;
-import io.swagger.v3.oas.models.OpenAPI;
-import io.swagger.v3.oas.models.info.Info;
-import io.swagger.v3.oas.models.info.License;
-import org.springframework.context.annotation.Bean;
-import org.springframework.context.annotation.Configuration;
-
-@Configuration
-public class OpenAPIConfiguration {
-
-    @Bean
-    public OpenAPI openAPI() {
-        return new OpenAPI()
-            .info(new Info().title("fis-hmc-api")
-                      .description("Create & update case and Upload and delete docs ")
-                      .version("v0.0.1")
-                      .license(new License().name("MIT").url("https://opensource.org/licenses/MIT")))
-            .externalDocs(new ExternalDocumentation()
-                              .description("README")
-                              .url("https://github.com/hmcts/spring-boot-template"));
-    }
-
-}
